+++ src/server/backend/mod.rs
@@ -14,7 +14,7 @@ pub trait InputSource {
 
 pub trait AudioEndpoint {
 	fn send_message(&self, message: Message) -> Result<(), String>;
-	fn stream_info(&self) -> (u8, u16);
+	fn stream_info(&self) -> (u8, u32);
 	fn exit(&self);
 }
